Refactor IOT tests into dedicated test files
